- handle Ns_SockState masks with explicit casts (to unsigned int) for   bit operations - more adjustments to new enums - reduce implicit conversions - name handled enum values explicitly in switch statements - avoid returns in switch statements - prefer return type bool with (NS_TRUE/NS_FALSE) over return type int (with  0 / 1) - add more comments to sockcallback.c

+++ nsd/event.c
@@ -116,8 +116,8 @@ Ns_CreateEventQueue(int maxevents)
 
     assert(maxevents > 0);
 
-    queuePtr = ns_calloc(1U, sizeof(EventQueue) + (sizeof(Event) * maxevents));
-    queuePtr->pfds = ns_calloc(maxevents + 1, sizeof(struct pollfd));
+    queuePtr = ns_calloc(1u, sizeof(EventQueue) + (sizeof(Event) * (size_t)maxevents));
+    queuePtr->pfds = ns_calloc((size_t)maxevents + 1u, sizeof(struct pollfd));
     if (ns_sockpair(queuePtr->trigger) != 0) {
         Ns_Fatal("taskqueue: ns_sockpair() failed: %s",
                  ns_sockstrerror(ns_sockerrno));
@@ -223,7 +223,7 @@ Ns_EventCallback(Ns_Event *event, Ns_SockState when, const Ns_Time *timeoutPtr)
      * Add to the waiting list if there are events or a timeout.
      */
 
-    if (evPtr->events != 0U || timeoutPtr != NULL) {
+    if (evPtr->events != 0 || timeoutPtr != NULL) {
         evPtr->status = EVENT_WAIT;
     } else {
         evPtr->status = EVENT_DONE;
